requirements.txt
discord.py
pyrebase4
Cybernator